+++ tests/run_tests.js
@@ -1,4 +1,4 @@
 const {run_tests} = require("./test_framework.js")
-require("./frontend/submit_request.js")
+require("./frontend/request.js")
 
 run_tests()
